Set quantity fields as Floats rather than BigDecimal
Decimal fields will become Floats by default where as converting to a
BigDecimal requires an extra operation. While the impact is small, it
will generally be unnecessary and, as such, we can leave it up to the
consumer to decide if they need a BigDecimal

+++ lib/dear_inventory/models/purchases/stock_line.rb
@@ -14,7 +14,7 @@ module DearInventory
           },
           Quantity: {
             name: :quantity,
-            type: :BigDecimal,
+            type: :Float,
           },
           ProductID: {
             name: :product_id,
